An Ansible Role to setup a Tezos node in Docker
This Ansible Role aims to make it easy and fast for an Ansible user to
setup a new Tezos Node.

The role is heavily parameterised, allowing users to deploy nodes for
different Tezos networks (mainnet/carthagenet/labnet/etc..) and various
economic protocols to support block transitions.

It supports different bootstrap strategies, namely syncing from genesis
or importing a snapshot for fast bootstrapping.

This is a first release, so we anticipate some rouge edges. File your
hardships and woes in the form of Issues on the projects github issue
tracker and we shall endeavour to liberate you from your woes and
hardships (within the narrow scope of this Ansible Role, we, alas, are
willing but unable to assist you with your in-laws, neck-pain,
pets-propensity-to-obsessively-lick-despite-your-earnest-pleas-to-said-pet,
or that rash).

Go forth and Tezos.