Relocate tezos-client config to the default location
Previously the role put --base-dir in /data/client which in hindsight is
weird. The role now mounts the client directory to the conventional
location /home/tezos/.tezos-client

The host location is the same `/srv/tezos/mainnet_client`.

We now also initialize the tezos-client config with the hostname of the
node. When a user attaches to a container, the `tezos-client` command
"just works" without specifying any additional flags.